+++ DiverseSelector/distance.py
@@ -92,28 +92,6 @@ def compute_distance_matrix(features: np.ndarray,
     return dist
 
 
-def distance_to_similarity(x: np.ndarray, dist: bool = True) -> np.ndarray:
-    """Convert between distance and similarity matrix.
-
-    Parameters
-    ----------
-    x : ndarray
-        Symmetric distance or similarity array.
-    dist : bool
-        Confirms the matrix is distance.
-
-    Returns
-    -------
-    y : ndarray
-        Symmetric distance or similarity array.
-    """
-    if dist is True:
-        y = 1 / (1 + x)
-    else:
-        y = (1 / x) - 1
-    return y
-
-
 def pairwise_similarity_bit(feature: np.array, metric: str) -> np.ndarray:
     """Compute the pairwaise similarity coefficients.
